Exceptions treatment

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,19 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AppError } from "@utils/AppError";
+
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroupName: string) {
     try {
         const storedGroups = await groupsGetAll();
+
+        const alreadyRegisteredGroup = storedGroups.includes(newGroupName);
+
+        if (alreadyRegisteredGroup) {
+            throw new AppError('Group already exists');
+        }
+
         const storage = JSON.stringify([...storedGroups, newGroupName])
         await AsyncStorage.setItem(GROUP_COLLECTION, storage);
     } catch (error) {
